blog_s2_v1

+++ customer/views.py
@@ -4,25 +4,46 @@ from .models import Customer, Contact
 from .form import CustomerForm, ContactForm
 from django.db.models import Q
 from django.contrib import messages
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from datetime import date
 
 def index(request):
-    customers = Customer.objects.all()
-
+    if request.user.is_staff == False:
+        return redirect('/home/')
+    
+    customers_list = Customer.objects.all()   
+    
     query = request.GET.get('q')
     if query:
-        customers = customers.filter(
+        customers_list = customers_list.filter(
             Q(name__icontains=query) |
             Q(address__icontains=query) |
-            Q(phone__icontains=query)             
+            Q(phone__icontains=query) |
+            Q(contacts__name__icontains=query) | 
+            Q(contacts__email__icontains=query) |
+            Q(contacts__phone__icontains=query)                                  
             ).distinct()
         
-
+    paginator = Paginator(customers_list, 5) # Show 25 contacts per page
+    page = request.GET.get('sayfa')
+    try:
+        customers = paginator.page(page)
+    except PageNotAnInteger:        
+        customers = paginator.page(1) # If page is not an integer, deliver first page.
+    except EmptyPage:        
+        customers = paginator.page(paginator.num_pages)    # If page is out of range (e.g. 9999), deliver last page of results.
+        
+    
     context = {
         'customers': customers,
+        
     }
     return render(request, 'customer/index.html', context)
 
 def create(request):
+    if request.user.is_staff == False:
+        return redirect('/home/')
+    
     form = CustomerForm(request.POST or None)
     if form.is_valid():
         form.save()
@@ -33,6 +54,9 @@ def create(request):
     return render(request, 'customer/form.html', context)
 
 def detail(request, pk):
+    if request.user.is_staff == False:
+        return redirect('/home/')
+    
     customer = get_object_or_404(Customer, pk=pk)
 
     form = ContactForm(request.POST or None)
@@ -55,6 +79,9 @@ def detail(request, pk):
     return render(request, 'customer/detail.html', context)
 
 def update(request, pk):
+    if request.user.is_staff == False:
+        return redirect('/home/')
+    
     customer = get_object_or_404(Customer, pk=pk)
     form = CustomerForm(request.POST or None, instance=customer)
     if form.is_valid():
@@ -66,12 +93,18 @@ def update(request, pk):
     return render(request, 'customer/form.html', context)
 
 def delete(request, pk):
+    if request.user.is_staff == False:
+        return redirect('/home/')
+    
     customer = get_object_or_404(Customer, pk=pk)
     customer.delete()
     return redirect("customer:index")
 
 
 def contact_update(request, pk):
+    if request.user.is_staff == False:
+        return redirect('/home/')
+    
     contact = get_object_or_404(Contact, pk=pk)    
     customer= get_object_or_404(Customer, id = contact.customer_id)
 
@@ -89,6 +122,9 @@ def contact_update(request, pk):
     return render(request, 'post/form.html', context)
 
 def contact_delete(request, pk):
+    if request.user.is_staff == False:
+        return redirect('/home/')
+    
     contact = get_object_or_404(Contact, pk=pk)
     customer = get_object_or_404(Customer, id = contact.customer_id)
     if request.user.is_authenticated == False:
